+++ NEWS.md
@@ -1,6 +1,7 @@
 # patRoon 0.1.0.9000
 
 ## November 2018
+* Changed `nintersects` default for plotUpSet so that all intersections are plotted by default.
 * plotChord() now properly stops if nothing overlaps.
 * replicateGroupSubtract() now removes replicate groups that were subtracted.
 * Fixed: replicateGroupSubtract() now correctly takes maximum mean intensity for threshold determination when multiple rGroups are specified.